requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-dotenv>=1.0
requests>=2.31
streamlit>=1.31
openai>=1.40
kubernetes>=29.0
